Add files via upload
I'm uploading this now as I seek to move on to studying frontend further in order to improve myself. I feel I've done my best for this project and will come back to it once I've improved. 

My to-dos:
-sync audio with mobile safari
-do more code refactoring
-add more games
-make page load faster

this is a row, but playable version of the simon game